requirements.txt
cycler==0.10.0
Cython==0.29.15
kiwisolver==1.1.0
lxml==4.5.0
matplotlib==3.2.0
numpy==1.18.1
Pillow==7.0.0
pip==19.2.3
pycocotools==2.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
setuptools==41.2.0
six==1.14.0
torch==1.4.0
